Updated CMakeLists and made use of smart pointers
OS X Mavericks changed the default standard library which programs are
linked against - the FBX SDK doesn't play nicely with it, and so the
CMakeLists.txt file now explicitly states the standard library to use.

With the lack of full C++11 support (FBX SDK does not like linking
against the incorrect standard library), shared_ptr and unique_ptr are
unavailable. The next best thing is auto_ptr, even if it is marked as
deprecated in C++11...